requirements.txt
transformers
torch
torchvision
accelerate
python-dotenv
openai
streamlit